Use Leader-with-lease leader election
The leader-for-life implementation blocks the whole operator, only
alloweing for one instance of both the controllre and the webhook to
run.

With the leader-with-lease implementation we can run the webhook on all
instances of the operator, while keeping the operator running with
leader election (meaning, only one instance runs the reconcile loop).