Merge selinux-policy-installer-operator
It used to be in a separate operator as a premature optimization
measure... but it didn't quite work that way. Once the operator creates
an object, the object type starts being tracked in the cache. So... This
pretty much invalidates my initial assumptions. I instead add the
controller that tracks configmaps and just ensure that we're tracking
the appropriate namespace in the reconciler.

This also adds the necessary bits to the role to be able to spawn
privileged containers.

On the other hand, the namespace was switched to have the -openshift
prefix.