Add "apply" key to SelinuxPolicy CR
This adds the "apply" key CR to the CR, which will indicate if we should
run the installation or not.

Note that currently you can only apply this once. So, if you set to true
apply, it'll be installed. If you set apply to false after that, nothing
will happen.