hash installer pod name
We used to have predictable names for the installer pod, however, due to
the podname being the name of the pod, the namespace, and node name,
they would usually get very long. This resulted in names needing to
be trimmed.

This was not ideal and ended up resulting in name conflicts. To address
this, we hash the name instead, and use the hash as a pod name. Given
the usual representation of hashes, this will still fulfil the need of
having DNS compliant names.