Make webhook skip important namespaces
The webhook needs to be able to skip the namespace of the
selinux-operator, if not, it'll deadlock. So we set the run-level to "1"
in order to avoid this, and a namespace selector for the webhook.

On the other hand, it does make sense for the webhook not to act on
other critical namespaces, so skips them too.